modifs

+++ TPVOL_JPA/src/main/java/sopra/vol/repository/IPassagerRepository.java
@@ -2,6 +2,6 @@ package sopra.vol.repository;
 
 import sopra.vol.model.Passager;
 
-public interface IPassagerRepository extends IRepository<Passager, Integer> {
+public interface IPassagerRepository extends IRepository<Passager, Long> {
 
 }
